requirements.txt
PyGithub>=2.2.0
tomli>=2.0.0; python_version < "3.11"
